backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
python-dotenv>=1.0
httpx>=0.27
pypdf>=4.0          # <-- ESTE FALTABA SEGURO
gunicorn>=21.0
Pillow>=10.0